Update error so code is API error code, not HTTP code
We use error classes to map to HTTP response codes that are considered
errors so we’d rather have the error code be the API error code.
Updated to parse that from body if given. Update tests to account for
changes. Fix changes to delete_card as now we’ll return the body rather
a boolean since we still raise an error so rather avoid that confusion.