Update verify method and rename to verify_card
Update required args. The paymentToken verification is for the case if
someone wants to save a card first and then verify it (seems odd to do
especially if card is invalid). Either way we’ll have a different named
method for that case to support it so we can make use of required args.
Updated tests.