Add camel case refinement and test
I’d like this library to be able to accept request payloads in snake
case but for that we’d need to convert camel case since that’s what the
Paysafe API accepts. I think its best to have this so that way when
using this Ruby gem in an app you don’t have mixed case just because
the API accepts data in a certain form.