Add Result object so responses are POROs rather than hashes
Instead of doing `response[:paymentToken]` we can now do
`response.payment_token` as we create a Result object from the response
we get back that generates getter methods (and predicates) for all
attribute keys. They are memoized. Further if a response hash has
nested hashes those too are converted to Result objects when retrieving
the value for the first time.

This is a large sweeping change where most client methods now convert
snake_case payloads to camelCase which is what the Paysafe REST API
expects.