Set record on default cassette options instead
We don't want to delete all cassettes to re-record. VCR has this capability baked in already which is better since we can override the default for tests that we know we can't re-record without additional updates (e.g. processing payments that require an involved manual UI flow).

While we still support the environment variable, this allows us to use in conjunction with specific tests too, not just the entire suite. So if we want to only run a single test file and re-record that, we can.