Remove VCR on/off helper methods
These are no longer necessary as the default is VCR is on and webmock as well. I have better understanding of how to use VCR so we prefer to have any new test cases use it rather than not. If a new test case is being introduced, its easy to have it re-record every time while working on a test by passing different options to the use_cassette method.