Update create/get payment tests
These were the result of new payment handle tokens now generated using Paysafe Checkout that had simulator set to external, so we are actually hitting the test environment of other third party services, in this case Sightline. Originally what we were hitting was a Paysafe mocked environment.

Updated the balance check in gateway response to use a regex match since the format varies here and in other endpoints which is odd. We don't really care about the value just that its there so a simple regex match that its a number (integer or decimal) is fine.