Remove code branches that don't apply
For snake case, we only call that on the API response body which we first parse the JSON with symbolize names on so the other branches won't ever apply and are reported as uncovered.

Same applies for camel case conversion since we use symbol based kwargs but even if its a string it will get converted as expected.

For both of these we could have something else other than a string or symbol as a key but that is not something the library even supports. We have no use case for that.

For the result class, the last branching condition of the `get_value_by` method wouldn't ever apply.

With these changes we should now be at 100% code coverage.